Fix: turmite returning wrong coordinates in step.
The coordinates returned as the result of an invocation of step were the
ones for the new turmite position, i.e. after the turmite advanced one
step. As a result the result the color change always occured at the
wrong position, which made the 'world' saved internally in each turmite
be different from the actual displayed image.
Using the old position in the return value should fix that.