Remove 'colors' array from turmite object.
The mapping from abstract colors to a concrete displayable
representation should not be done/saved inside each turmite. That's why
the colors array should not be a member of a turmite object and is
therefore removed with this patch.
Unfortunately this means that the whole conflict handling code has to
removed as now the color used to actually display a combined step is no
longer available in the corresponding function. But the conflict
handling code was wrong and has to be rethought anyway, so this is no
big deal.