Enable different color representations.
Turmites can now use different actual colors to represent their abstract
colors when drawing. This helps to distinguish which parts of the world
have last been affected by which turmite. This also means a table with
more than two colors has to be used for drawing.
If two turmites want to draw on the same position at the same time their
colors representations are averaged, but they still have to be
compatible in their abstract color representations, i.e. both turmites
have to agree on which abstract color should be put on the position.
Collisions are only handled on the representation level. Conflict
handling on the turmite level still has to be implemented.