Implement random, consistent turmite behaviour.
If a turmite encounters a color, state combination for which no move is
known, it chooses one move randomly. The move is cached, so when the
combination of color and state is encountered again, the same move as
before is executed. This is why the turmite behaviour is random but
consistent.
The move is controlled via the turmites colors and state members, each
specifying the maximum number of colors and states which may be used.