Make turmites meet in the center.
Having the image be 499px squared, sets the image's center to be at
pixel (249, 249) and the new turmite positions make them meet at exactly
that point, although the ones building horizontal and vertical highways
arrive there earlier than the ones building diagonal highways.
The turmites are also reordered to use the colors in a way which was
more to my liking.