Fit all elements in the same column to one width
It might look complicated, but all the code does is:

  * Take all elements which have the class `columns` attached to it
    (These are treated as elements containing columns.)
  * Look for all children in such an element which have a class attached
    to it that starts with `column-`.
  * These elements are then grouped by whatever comes after the
    `column-` prefix, because having the same suffix should mean that
    these elements belong to the same column.
    (Note: due to the step above, elements contained in different
     `.columns` containers are always treated as belonging to different
     columns.)
  * The widths of all the elements in one column are then set to the
    largest width a single element in the column has.

This is what `fit_columns` does, and it's also why it gets called once
the document's `ready` event is fired.
Last but not least, setting the `display` style of labels to
`inline-block` was necessary, as otherwise, the label's `width` style
would have no visible effect. Instead of directly targeting the labels,
I just set to `display` style for all elements comprising a column, just
in case.