Start using the default layout again
It might generate annoying stuff, but it looks better and I'm going to
try out other layouts anyway. Also, I want to have a commit which is
only on this branch to check whether I can use a branch that is
different from master to generate the GitHub Pages for this repository.